Control deck command
